Task3

+++ src/main/java/com/mindhub/homebanking/controllers/ClientController.java
@@ -1,7 +1,6 @@
 package com.mindhub.homebanking.controllers;
 
 import com.mindhub.homebanking.dtos.ClientDTO;
-import com.mindhub.homebanking.models.Client;
 import com.mindhub.homebanking.repositories.ClientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -24,7 +23,7 @@ public class ClientController {
 
     }
     @RequestMapping("/clients/{id}")
-    public ClientDTO getClient(@PathVariable Integer id){
+    public ClientDTO getClient(@PathVariable Long id){
         return clientRepository.findById(id).map(ClientDTO::new).orElse(null);
     }
 }
